Started refactoring wiring diagrams stuff
