+++ Project.toml
@@ -1,11 +1,12 @@
 name = "SemanticModels"
 uuid = "f5ac2a72-33c7-5caf-b863-f02fefdcf428"
-authors = ["James Fairbanks <james.fairbanks@gtri.gatech.edu>"]
+authors = ["James Fairbanks <james.fairbanks@gtri.gatech.edu>", "Micah Halter <micah.halter@gtri.gatech.edu>"]
 version = "0.2.0"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Cassette = "7057c7e9-c182-5462-911a-8362d720325c"
+Catlab = "134e5e36-593f-5add-ad60-77f754baafbe"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DataFramesMeta = "1313f7d8-7da2-5740-9ea0-a2ca25f37964"
